+++ src/Private/apiCalls/Invoke-DBPoolRequest.ps1
@@ -32,7 +32,7 @@ function Invoke-DBPoolRequest {
     .PARAMETER data
         Defines the data to be sent with the API request body when using POST or PATCH
 
-    .PARAMETER DBPool_JSON_Conversion_Depth
+    .PARAMETER jsonDepth
         Defines the depth of the JSON conversion for the 'data' parameter request body
 
     .PARAMETER allPages
@@ -99,14 +99,21 @@ function Invoke-DBPoolRequest {
 
         [Parameter(Mandatory = $false)]
         #[ValidateRange(0, [int]::MaxValue)]
-        [int]$DBPool_JSON_Conversion_Depth = 5,
+        [int]$jsonDepth = $DBPool_JSON_Conversion_Depth,
 
         [Parameter(DontShow = $true, Mandatory = $false)]
         [Switch]$allPages
 
     )
 
-    begin {}
+    begin {
+
+        if (-not $PSBoundParameters['jsonDepth'] -and $null -eq $DBPool_JSON_Conversion_Depth) {
+            $jsonDepth = 100
+            Write-Debug "The 'jsonDepth' parameter was not set. Using the default value of [ $jsonDepth ]"
+        }
+
+    }
 
     process {
 
@@ -126,7 +133,7 @@ function Invoke-DBPoolRequest {
         if ($null -eq $data) {
             $request_Body = $null
         } else {
-            $request_Body = $data | ConvertTo-Json -Depth $DBPool_JSON_Conversion_Depth
+            $request_Body = $data | ConvertTo-Json -Depth $jsonDepth
         }
 
         try {